backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.14
djangorestframework-simplejwt>=4.8,<5.0
django-cors-headers>=3.10.0
django-environ>=0.9.0
psycopg2==2.8.6
requests>=2.26
web3>=6.0
opencv-python>=4.5
tensorflow>=2.8.0
keras>=2.8.0
scipy>=1.8.0
pillow>=9.0.0
python-dotenv>=0.19.0


